feat(TransactionHistory): create TransactionHistory component
This commit adds a new component, TransactionHistory, to the project. This component is responsible for displaying details about a transaction. It will be used in the future to provide a detailed view of a user's transaction history.

fix(HistoryAccount): update GestorDector behavior and add header

This commit updates the HistoryAccount component to fix an issue with the GestorDector feature. Now, GestorDector will only work on the header of this component and will never close it. Additionally, a with a title and future filheader ter has been added, along with the TransationHistory component.